Parentable class into its own new module
* Parentable class into parentable module.
* delinking Parentable class needing to know about its own children.
